# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "اغتنم API Documentation",
        "description": "API documentation for اغتنم application",
        "contact": {
            "name": "Admin",
            "email": "admin@e8tanem.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get a specific lesson",
                "operationId": "546c214c4f0b15c37c5d46fc152b5c89",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID of lesson",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "subject_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Introduction to Calculus"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Learn the basics of calculus"
                                        },
                                        "video_url": {
                                            "type": "string"
                                        },
                                        "thumbnail_url": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "views_count": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "subject": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Mathematics"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Learn advanced mathematics"
                                                },
                                                "thumbnail": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Lesson not found or inactive"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lessons/{lesson}/stream": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Stream a specific lesson video",
                "description": "Stream the lesson video content with support for range requests and caching for mobile optimization",
                "operationId": "c8ac8f1b107a25597c39c3190aba6f94",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID of the lesson to stream",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Range",
                        "in": "header",
                        "description": "Range header for partial content requests (e.g., bytes=0-1000)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full video content stream",
                        "headers": {
                            "Content-Type": {
                                "description": "video/mp4",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Accept-Ranges": {
                                "description": "bytes",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Length": {
                                "description": "Size of the video in bytes",
                                "schema": {
                                    "type": "integer"
                                }
                            },
                            "Cache-Control": {
                                "description": "public, max-age=86400",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "206": {
                        "description": "Partial video content (for range requests)",
                        "headers": {
                            "Content-Type": {
                                "description": "video/mp4",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Range": {
                                "description": "bytes start-end/size",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Length": {
                                "description": "Size of the requested range in bytes",
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson not found, inactive, or video file not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}